Export and Import JS script from a file to another
