+++ index.html
@@ -93,7 +93,7 @@
 
         <!-- followig div will contain the dynamic projects list that is displayed with JS -->
         <div class="projectsContainer"></div>
-        
+
         <!-- followig div will contain the dynamic popup page that is displayed with JS -->
         <div class="popupPage"></div>
       </section>
@@ -193,8 +193,8 @@
 
       <img class="longindicator" src="assests/images/longindicator.svg" alt="Indicator" />
     </footer>
-    <script src="./JS/main.js"></script>
-    <script src="./JS/popupPage.js"></script>
+    <script src="./JS/menu.js"></script>
+    <script type="module" src="./JS/popupPage.js"></script>
     <script src="./JS/projects.js"></script>
     <script src="./JS/contactForm.js"></script>
   </body>
